lightning: split engine and writer from local.go
