+++ br/pkg/lightning/backend/local/iterator.go
@@ -55,7 +55,7 @@ type duplicateIter struct {
 	nextKey   []byte
 	err       error
 
-	engineFile     *File
+	engine         *Engine
 	keyAdapter     KeyAdapter
 	writeBatch     *pebble.Batch
 	writeBatchSize int64
@@ -100,7 +100,7 @@ func (d *duplicateIter) flush() {
 }
 
 func (d *duplicateIter) record(key []byte, val []byte) {
-	d.engineFile.Duplicates.Inc()
+	d.engine.Duplicates.Inc()
 	d.err = d.writeBatch.Set(key, val, nil)
 	if d.err != nil {
 		return
@@ -170,7 +170,7 @@ func (d *duplicateIter) OpType() sst.Pair_OP {
 
 var _ kv.Iter = &duplicateIter{}
 
-func newDuplicateIter(ctx context.Context, engineFile *File, opts *pebble.IterOptions) kv.Iter {
+func newDuplicateIter(ctx context.Context, engine *Engine, opts *pebble.IterOptions) kv.Iter {
 	newOpts := &pebble.IterOptions{TableFilter: opts.TableFilter}
 	if len(opts.LowerBound) > 0 {
 		newOpts.LowerBound = codec.EncodeBytes(nil, opts.LowerBound)
@@ -179,27 +179,27 @@ func newDuplicateIter(ctx context.Context, engineFile *File, opts *pebble.IterOp
 		newOpts.UpperBound = codec.EncodeBytes(nil, opts.UpperBound)
 	}
 	logger := log.With(
-		zap.String("table", common.UniqueTable(engineFile.tableInfo.DB, engineFile.tableInfo.Name)),
-		zap.Int64("tableID", engineFile.tableInfo.ID),
-		zap.Stringer("engineUUID", engineFile.UUID))
+		zap.String("table", common.UniqueTable(engine.tableInfo.DB, engine.tableInfo.Name)),
+		zap.Int64("tableID", engine.tableInfo.ID),
+		zap.Stringer("engineUUID", engine.UUID))
 	return &duplicateIter{
 		ctx:        ctx,
-		iter:       engineFile.db.NewIter(newOpts),
-		engineFile: engineFile,
-		keyAdapter: engineFile.keyAdapter,
-		writeBatch: engineFile.duplicateDB.NewBatch(),
+		iter:       engine.db.NewIter(newOpts),
+		engine:     engine,
+		keyAdapter: engine.keyAdapter,
+		writeBatch: engine.duplicateDB.NewBatch(),
 		logger:     logger,
 	}
 }
 
-func newKeyIter(ctx context.Context, engineFile *File, opts *pebble.IterOptions) kv.Iter {
+func newKVIter(ctx context.Context, engine *Engine, opts *pebble.IterOptions) kv.Iter {
 	if bytes.Compare(opts.LowerBound, normalIterStartKey) < 0 {
 		newOpts := *opts
 		newOpts.LowerBound = normalIterStartKey
 		opts = &newOpts
 	}
-	if !engineFile.duplicateDetection {
-		return pebbleIter{Iterator: engineFile.db.NewIter(opts)}
+	if !engine.duplicateDetection {
+		return pebbleIter{Iterator: engine.db.NewIter(opts)}
 	}
-	return newDuplicateIter(ctx, engineFile, opts)
+	return newDuplicateIter(ctx, engine, opts)
 }
